hdf5 writer internals improved.
